refactor(icon): cleanup and storybook made prettier
